Create default_context_with_storage() for tests
