+++ crates/networking/rpc/eth/gas_price.rs
@@ -51,59 +51,24 @@ impl RpcHandler for GasPrice {
 mod tests {
     use super::GasPrice;
     use crate::eth::test_utils::{
-        add_eip1559_tx_blocks, add_legacy_tx_blocks, add_mixed_tx_blocks, setup_store,
-        BASE_PRICE_IN_WEI,
+        add_eip1559_tx_blocks, add_legacy_tx_blocks, add_mixed_tx_blocks,
+        default_context_with_storage, setup_store, BASE_PRICE_IN_WEI,
     };
 
-    use crate::utils::test_utils::example_local_node_record;
     use crate::{
-        rpc::{map_http_requests, RpcApiContext, RpcHandler},
-        utils::{parse_json_hex, test_utils::example_p2p_node, RpcRequest},
+        rpc::{map_http_requests, RpcHandler},
+        utils::{parse_json_hex, RpcRequest},
     };
     #[cfg(feature = "based")]
     use crate::{EngineClient, EthClient};
     #[cfg(feature = "based")]
     use bytes::Bytes;
-    use ethrex_blockchain::Blockchain;
-    use ethrex_p2p::sync_manager::SyncManager;
-    #[cfg(feature = "l2")]
-    use ethrex_storage_l2::{EngineTypeL2, StoreL2};
-    #[cfg(feature = "l2")]
-    use secp256k1::{rand, SecretKey};
     use serde_json::json;
-    use std::sync::Arc;
-
-    async fn default_context() -> RpcApiContext {
-        let storage = setup_store().await;
-
-        let blockchain = Arc::new(Blockchain::default_with_store(storage.clone()));
-        RpcApiContext {
-            storage,
-            blockchain,
-            jwt_secret: Default::default(),
-            local_p2p_node: example_p2p_node(),
-            local_node_record: example_local_node_record(),
-            active_filters: Default::default(),
-            syncer: Arc::new(SyncManager::dummy()),
-            #[cfg(feature = "based")]
-            gateway_eth_client: EthClient::new(""),
-            #[cfg(feature = "based")]
-            gateway_auth_client: EngineClient::new("", Bytes::default()),
-            #[cfg(feature = "based")]
-            gateway_pubkey: Default::default(),
-            #[cfg(feature = "l2")]
-            valid_delegation_addresses: Vec::new(),
-            #[cfg(feature = "l2")]
-            sponsor_pk: SecretKey::new(&mut rand::thread_rng()),
-            #[cfg(feature = "l2")]
-            l2_store: StoreL2::new("test-store", EngineTypeL2::InMemory)
-                .expect("Fail to create in-memory db test"),
-        }
-    }
 
     #[tokio::test]
     async fn test_for_legacy_txs() {
-        let context = default_context().await;
+        let storage = setup_store().await;
+        let context = default_context_with_storage(storage).await;
 
         add_legacy_tx_blocks(&context.storage, 100, 10).await;
 
@@ -115,7 +80,8 @@ mod tests {
 
     #[tokio::test]
     async fn test_for_eip_1559_txs() {
-        let context = default_context().await;
+        let storage = setup_store().await;
+        let context = default_context_with_storage(storage).await;
 
         add_eip1559_tx_blocks(&context.storage, 100, 10).await;
 
@@ -126,7 +92,8 @@ mod tests {
     }
     #[tokio::test]
     async fn test_with_mixed_transactions() {
-        let context = default_context().await;
+        let storage = setup_store().await;
+        let context = default_context_with_storage(storage).await;
 
         add_mixed_tx_blocks(&context.storage, 100, 10).await;
 
@@ -137,7 +104,8 @@ mod tests {
     }
     #[tokio::test]
     async fn test_with_not_enough_blocks_or_transactions() {
-        let context = default_context().await;
+        let storage = setup_store().await;
+        let context = default_context_with_storage(storage).await;
 
         add_mixed_tx_blocks(&context.storage, 100, 0).await;
 
@@ -148,7 +116,8 @@ mod tests {
     }
     #[tokio::test]
     async fn test_with_no_blocks_but_genesis() {
-        let context = default_context().await;
+        let storage = setup_store().await;
+        let context = default_context_with_storage(storage).await;
         let gas_price = GasPrice {};
         // genesis base fee is = BASE_PRICE_IN_WEI
         let expected_gas_price = BASE_PRICE_IN_WEI;
@@ -166,8 +135,8 @@ mod tests {
         });
         let expected_response = json!("0x3b9aca00");
         let request: RpcRequest = serde_json::from_value(raw_json).expect("Test json is not valid");
-        let mut context = default_context().await;
-        context.local_p2p_node = example_p2p_node();
+        let storage = setup_store().await;
+        let context = default_context_with_storage(storage).await;
 
         add_legacy_tx_blocks(&context.storage, 100, 1).await;
